backend/requirements.txt
# Core App Requirements
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
pydantic_core==2.33.2
starlette==0.46.2
anyio==4.9.0
click==8.2.0
h11==0.16.0
idna==3.10
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2

# NLP & Transformers
transformers==4.41.1         # Used for emotion detection and topic analysis
torch==2.3.0                 # Required backend for transformers (used under the hood)
sentence-transformers==2.7.0 # For embedding similarity and advanced NLP

# Utility
requests==2.31.0             # For HTTP calls to APIs (e.g. Gemini if used externally)
httpx==0.27.0                # Async HTTP if needed
python-dotenv==1.0.1         # To load .env config
scikit-learn==1.4.2
nltk==3.8.1
spacy==3.7.4
motor==X.Y.Z # Use a specific version, e.g., motor==4.0.0
pymongo==A.B.C # Use a specific version, e.g., pymongo==4.7.0
